moving FamingNonGold Models to VillageModels folder
